amazon-seller-automation/requirements.txt
python-amazon-sp-api
gspread
google-auth
python-dotenv
